requirements.txt
numpy
torch==1.8.1
matplotlib
Pillow
torchvision
scipy
